[REFACTOR] DB 구조 변경
- 결제 결과 관리를 위한 PaymentHistory Entity 추가
- 파일 순서 관리 위한 FilesSequence Entity 추가
- 사용되지 않는 Funding Entity 내 saveStatus 컬럼 삭제
