+++ src/main/java/com/kosa/backend/funding/project/entity/Funding.java
@@ -3,6 +3,7 @@ package com.kosa.backend.funding.project.entity;
 import com.kosa.backend.common.entity.Auditable;
 import com.kosa.backend.funding.project.entity.enums.MakerType;
 import com.kosa.backend.funding.support.entity.FundingSupport;
+import com.kosa.backend.payment.entity.PaymentHistory;
 import com.kosa.backend.user.entity.Maker;
 import jakarta.persistence.*;
 import lombok.*;
@@ -56,9 +57,6 @@ public class Funding extends Auditable {
     private String fundingTag;
 
     @Column(nullable = false)
-    private boolean saveStatus;
-
-    @Column(nullable = false)
     private boolean isPublished;
 
     private LocalDateTime publishDate;
@@ -85,4 +83,7 @@ public class Funding extends Auditable {
     @ManyToOne
     @JoinColumn(name = "funding_support_id", nullable = true)
     private FundingSupport fundingSupport;
+
+    @OneToMany(mappedBy = "funding", cascade = CascadeType.ALL)
+    private List<PaymentHistory> paymentHistory;
 }
