+++ src/main/java/com/kosa/backend/payment/entity/Payment.java
@@ -1,12 +1,13 @@
 package com.kosa.backend.payment.entity;
 
 import com.kosa.backend.common.entity.Auditable;
-import com.kosa.backend.payment.entity.eunms.PaymentStatus;
+import com.kosa.backend.payment.entity.enums.PaymentStatus;
 import com.kosa.backend.user.entity.User;
 import jakarta.persistence.*;
 import lombok.Setter;
 
 import java.time.LocalDateTime;
+import java.util.List;
 
 @Setter
 @Entity
@@ -41,4 +42,7 @@ public class Payment extends Auditable {
     @ManyToOne
     @JoinColumn(name = "coupon_id")
     private Coupon coupon;
+
+    @OneToMany(mappedBy = "payment", cascade = CascadeType.ALL)
+    private List<PaymentHistory> paymentHistory;
 }
