+++ src/main/java/com/kosa/backend/payment/entity/PaymentHistory.java
@@ -1,0 +1,26 @@
+package com.kosa.backend.payment.entity;
+
+import com.kosa.backend.common.entity.Auditable;
+import com.kosa.backend.funding.project.entity.Funding;
+import jakarta.persistence.*;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+
+@EqualsAndHashCode(callSuper = true)
+@Data
+@Entity
+@Table(name = "PAYMENT_HISTORY")
+public class PaymentHistory extends Auditable {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private int id;
+
+    @ManyToOne
+    @JoinColumn(name = "payment_id", nullable = false)
+    private Payment payment;
+
+    @ManyToOne
+    @JoinColumn(name = "funding_id", nullable = false)
+    private Funding funding;
+}
