+++ src/main/java/com/kosa/backend/payment/repository/PaymentHistoryRepository.java
@@ -1,0 +1,7 @@
+package com.kosa.backend.payment.repository;
+
+import com.kosa.backend.payment.entity.PaymentHistory;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PaymentHistoryRepository extends JpaRepository<PaymentHistory, Integer> {
+}
